requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.15.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tzdata==2025.1
